Core/C#: Finished scripts for DataGridView.cs and take result from excel files.

+++ Projects/C#/Predfakturace/Predfakturace/UniqueValue.cs
@@ -9,7 +9,7 @@ namespace Predfakturace
     static class UniqueValue
     {
         public static List<string> traceToFile = new List<string>();
-        public static List<int> money = new List<int>();
+        public static List<double> money = new List<double>();
         public static string finalyPath;
     }
 }
